globalsByName {globals}	R Documentation

Locates and retrieves a set of global variables by their names

Description

Locates and retrieves a set of global variables by their names

Usage

globalsByName(names, envir = parent.frame(), mustExist = TRUE, ...)

Arguments

names
A character vector of global variable names.

envir
The environment from where to search for globals.

mustExist
If TRUE, an error is thrown if the object of the identified global cannot be located. Otherwise, the global is not returned.

...
Not used.

Value

A Globals object.

[Package globals version 0.14.0 Index]